improve chaining of pod selectors
