+++ pkg/collectors/instance_select.go
@@ -28,50 +28,3 @@ func LifecycleTriggeredOlderThan(age time.Duration) Selector {
 func LifecycleDeleted(i *Instance) bool {
 	return i.ASG.Deleted
 }
-
-// InstanceQuery returns a dummy selector that selects all instances. It is
-// used to make chaining selectors prettier while making sure the type is
-// correct.
-//
-// Without:
-//    Selector(HasEC2Data).
-//        Select(HasASGData).
-//        Filter(LifecycleDeleted)
-// With:
-//    InstanceQuery().
-//        Select(HasEC2Data).
-//        Select(HasASGData).
-//        Filter(LifecycleDeleted)
-func InstanceQuery() Selector {
-	return func(i *Instance) bool {
-		return true
-	}
-}
-
-func (s1 Selector) Select(s2 Selector) Selector {
-	return Selector(func(i *Instance) bool {
-		return s1(i) && s2(i)
-	})
-}
-
-func (s1 Selector) Filter(s2 Selector) Selector {
-	return Selector(func(i *Instance) bool {
-		return s1(i) && !s2(i)
-	})
-}
-
-func (is Selector) FilterByAllPods(ps PodSelector) Selector {
-	return Selector(func(i *Instance) bool {
-		if !is(i) {
-			return false
-		}
-
-		for _, pod := range i.Pods {
-			if !ps(&pod) {
-				return false
-			}
-		}
-
-		return true
-	})
-}
